requirements.txt
distlib==0.3.9
filelock==3.16.1
platformdirs==4.3.6
python-dotenv==1.0.1
requests~=2.32.3